+++ extensions/quick-git/src/quick-git.tsx
@@ -2,7 +2,7 @@ import { useLocalStorage } from "@raycast/utils"
 import { GitStatus } from "./components/GitStatus/GitStatus.js"
 
 export default function Command() {
-	const { value, isLoading } = useLocalStorage<string>("repo")
+	const { value, isLoading } = useLocalStorage<string>("selectedRepo")
 
 	return <GitStatus repo={value} isLoadingRepo={isLoading} />
 }